// src/modules/billing/rewards/types/reward-type.enum.ts
export enum RewardTypeEnum {
  DAILY_LOGIN = 'DAILY_LOGIN',
  POST_VIDEO_REWARD = 'POST_VIDEO_REWARD',
  POST_PHOTO_REWARD = 'POST_PHOTO_REWARD',
  CONTEST_PARTICIPATION = 'CONTEST_PARTICIPATION',
  REGISTRATION_REWARD = 'REGISTRATION_REWARD',
  RATE_APP = 'RATE_APP',
  LIKE_EARN = 'LIKE_EARN',
  LIKE_SPEND = 'LIKE_SPEND',
  SHARE_REWARD = 'SHARE_REWARD',
  SHARE_YEPS = 'SHARE_YEPS',
  PAYMENT_5000 = 'PAYMENT_5000',
  PAYMENT_15000 = 'PAYMENT_15000',
  PAYMENT_30000 = 'PAYMENT_30000',
  REFERRAL_REWARD = 'REFERRAL_REWARD',
  REGISTRATION_BONUS = 'REGISTRATION_BONUS',
}